Fix error in TMDB account movie api
Change parameter name for favorite call which was mistaken (thank you copy+paste...).
Change the value of http success response for favorite and watch list, 200 is specified in
the documentation but it is returning 201.
Change using selected state to activated state to control the favorite / watchlist button state,
so it will properly use the selector associated.