Create AuthenticatedApi
Similar to PublicApi, create a new class that servers to extracts the requests logic from MovieDetailActivity
that require to be authenticated with TMDB to be able to use, such as adding movies to your account watch list
or favorite them.

Note that they have additional dependencies in the api methods as the dependencies to authenticate (account id
and session token) as the services are still being built inside.