Extract account movie service as class field
Similar to the public api the service and it's request interceptor is extracted as a class field allowing
it to be reused among all the different methods that it declares. However as it depends on the authentication
information to be built, this parameters have also been extracted to the class context instead of being
introduced as parameters in the method call.