Extract dependencies from PublicApi
PublicApi requires the TMDB services to work, however its aim is to wrap the specific parameters that the
service needs to work, from the actual data that it requires (for example the movie id it is going retrieve).
It should not have to worry about creating the services, rather they should be provided to it.