Removed unused dependency in PublicApi
Before PublicApi had the responsibility of creating it's dependencies, forcing it to also have to handle
the creation of the dependencies the services had. As the services are now provided the PublicApi class
relied on clases that it really did not need, and have been removed in this commit.
This was because the class had dependencies that were not their own.